Prevent conflicting coordinate system ids
Since the next coordinate system id was stored on the state, this value
was not being propagated back up to siblings. The TransformUpdateState
is meant to only be updated downward. Pull out the next
CoordinateSystemId as another mutable argument in order to have it be
persistent through the entire tree traversal. There is no test for
this, but the previous design was obviously wrong.