Add a compact profiler mode.
This allows viewing a minimal amount of the most critical data from
the profiler, as shown in the screenshot below.

This is useful for a couple of reasons, namely:
 * Makes it easier to interact with the underlying page.
 * Reduces the amount of GPU time spent drawing the profiler itself.

The profiler code in general is a bit messy - this includes a few
hacks to work around the borrow checker. We should restructure this
a bit in the future.