Add a function to compute various type flags of a matrix.
This provides a more unified way to calculate various properties
of a matrix, rather than has_perspective_component etc.

The preserves_2d_alignment method returns true for transforms
with a 90 degree rotation. This function allows us to differentiate
between that and a pure translation matrix, which will be useful
for some of the upcoming clip / segment code.