Make incompatible coordinate system detection more conservative.
Only consider references frames with simple translations to be
compatible coordinate systems. This fixes some reftests in servo,
related to rotation transforms that do preserve 2d axis condition,
but cause problems (e.g. a 90 degree rotation).

In the future, we may be able to relax this and handle some more
complex cases, but these are edge cases anyway, so shouldn't have
a significant performance impact.