Remove RG8 and RGB8 support from texture cache.
RGB8 support is removed completely from WR. It's not supported
by many modern GPUs as a native format. Instead, expect the client
code to convert it to RGBx first.

RG8 support is removed from the texture cache, but remains as a
supported image format. It's used by some variants of YUV formats.
This allows RG8 images to be used as external images, but not
placed into the texture cache.