Only render in wrench on events, rather than spinning CPU at 100%.
This removes the old performance timing code (which is now unused),
and switches the event loop to (a) block until there are events
present and (b) consume all events that have been queued and are
available before drawing a frame.

Before this change, the event queue was getting swamped with
Awaken messages and not processing any scroll events (which our
glutin version only seems to provide if you call poll_events).

This will allow us to add support for zoom to wrench, which we
need for testing a variety of bugs that only occur under certain
zoom and DPI conditions. A side benefit is that wrench no longer
keeps the CPU active 100% of the time.

I've tested that binary replays still work, and that the reftests
still work with this change.