Update reftest images for #2262.
It's not clear why, but these reftests are failing for me on master
on my local machine. This could either be (a) my machine generating
different images due to a different freetype or (b) something weird
on CI where a test is failing and still being merged.

Opening this as a PR to see what CI makes of it...