Add support for regions to the segment builder.
This allows specifying an inner rect for the segment builder,
which defines the main primitive as a 9-patch. Then, when
segments are created, the region they belong to is passed to
the segment closure.

This allows us to do things like build segments for a box-shadow
primitive, where the box-shadow is made up of sub-regions (i.e.
the corners, edges and inner region are separate regions).

In the case of a fast box shadow path, the region can be used
to identify the UV sub-rect from the blurred image to draw with.