Remove external scroll ids from ClipId
Move the concept of external scroll ids to ScrollFrames themselves. It
was probably a mistake to integrate this into ClipId in the first place.
Moving it out allows us to simplify a few code paths and to remove user
defined ClipIds entirely, making things simpler and more consistent.
This will also allow fixing Servo bugs 19287, 17176 and 19648.