Fix for race condition when caching render tasks.
When the renderer receives multiple Frames from the backend
in a single render loop, it drops the old frames and just
draws the most recent frame.

If that old frame contained a render task to write to the
texture cache, then the texture cache will never be updated.

This patches tracks whether a frame has any texture cache
render tasks, and whether it has been drawn before. If a
frame which meets those conditions is being replaced, then
do a render of that frame first, skipping the main framebuffer
pass.

This is not ideal since we may end up drawing offscreen render
tasks that were batched but which aren't strictly necessary
for the texture cache render target. However, it is very rare
that the render backend is (a) producing frames faster than they
are being consumed by the renderer and (b) also contain texture
cache tasks.