Make WR hit-testing produce consistent results.
Ideally we would want the hit-test to operate against the previously
rendered frame. However, once a SetDisplayList call happens, the clip
scroll tree for the previouly rendered frame is clobbered, and so the
hit-test operates on inconsistent data and therefore produces
inconsistent results. This patch is a temporary workaround that triggers
a new render if needed in this scenario, so that at least the hit-test
produces consistent results.

Workaround for #2099