Disable shadow-ordering on mac.
This test is flaky, depending on the layout of the texture cache.

It's being affected by unrelated tests that run previous in CI.

What it's testing is platform independent, and covered by running it
on Linux (where it passes), so we can disable on mac for now until
we find the root cause.