Move hit-test commands to be processed after document rendering.
This adds a new field to DocumentOp, which contains "queries" that
update_document is meant to run after doing updates to the document.
HitTest commands now just trigger a render (if needed) but defer the
real work into a query. This ensures the hit-test is still run after the
document is rendered, but is more in line with the structure of the
current code, and does a better job of not violating implicit
assumptions about document rendering/publishing.

This fixes bustage from #2333.