Make hit testing consistent between frames
Move hit testing data structures to the Document. This make hit testing
results consistent with the currently rendered frame. To do this we are
required to copy important information from the ClipScrollTree during
the final step of producing a RenderedDocument, so this implies a
certain per-frame overhead. In the future we can look to make this less
expensive.

Fixes #2099.