Add hit testing support for ClipChains from the API
Hit testing now has support for tests against items that were clipped by
API-defined ClipChains. Both clipping ClipScrollNodes nodes and
ClipChains are converted to pairs of ClipScrollNodes and ClipChains for
purposes of the HitTester. This more-or-less matches the logical next
step in ClipScrollTree design and makes the logic for HitTesting
simpler.

Fixes #2383.