Port ps_blend to be a brush shader.
Previously, ps_blend applied the blend operation
in 2D screen-space. Now, blends are performed as
a normal brush shader. This provides several benefits:

* For a transformed primitive, we blend fewer pixels,
  since we only draw the composite step as a transformed
  primitive, rather than a 2D screen aligned rect.

* We can optionally apply all the usual features
  that are part of the brush shader, specifically
  clip masks, edge AA and pixel snapping.

* This lays the groundwork for supporting picture
  targets that are selectable to be either rasterized
  in local-space or screen-space, depending on what
  the caller requests as a performance / quality
  trade-off.

Next, we'll port hw_composite and composite shaders
to use the same brush format, and add support for
applying the source texture based on whether it was
rasterized in local or screen space.