Assign all ClipIds during display list construction
This change makes it so that all ClipIds are assigned during display
list construction. This will allow us to count the number of ClipIds in
each display list during construction, making it possible to more easily
map them onto simple array indices. The final goal of this work is to
eliminate the performance cost of hashing when flattening and rendering
display lists.