Use mipmaps and trilinear filtering for large, standalone textures.
Images that are large (currently this means one dimension > 512) are
placed into standalone textures rather than allocated into the
shared texture cache pages. Enabling mipmaps and trilinear filtering
on this subset of images is very simple, and also fixes the majority
of images where lack of mipmaps are a noticeable quality issue.

In the future, we can expand this to support mipmaps within the
shared cache pages, if desired, by manually drawing the mips with
a shader and making use of the mip-bias parameter to disable mip
selection for shared cache regions without mips.

Fixes #2023.