Only apply mip-maps if both image dimensions are > 512.
This is a hack / workaround for some test failures in gecko.

There are some test images with resolution 2048x32. These were
having mips enabled, but then failing reftests since the tests
were expecting bilinear filtering only. When these images were
being drawn with a size of 4x4 pixels, the smallest mip was
being selected, which is not what the test expects.

This workaround is a reasonable tradeoff for now - we get mip
mapping on large images, while allowing the existing tests to
pass.

In the future, we could consider some alternatives (such as a better
heuristic for enabling mips, or a different mip algorithm).