Pack extra user data value into BrushInstance structure.
In order to port the YUV image shader, we will need three
user data fields (one per channel).

To enable this work, we now pack the two render task
addresses for the picture and clip task as 2x u16 into
a single u32. This is all a bit messy still, but once
we have everything converted over to use the brush vertex
format, we can define it correctly as u16 attributes and
avoid all the packing / unpacking.