Convert ClipIds for Clips into indices in the ClipScrollTree
During display list flattening, we convert ClipIds for the clipping node
into ClipChainIndex (as well as allocating space for them in the
ClipChain array in the ClipScrollTree). This allows us to avoid doing a
HashMap lookup for the ClipChain for every primitive run. It also moves
us a little closer to a universal representation of ClipChains (both
API-defined ClipChains and ones derived from the hierarchy of the
ClipScroll tree).

This causes the amount of time spend doing HashMap lookups to decrease
in the profile, though they are still done for the positioning node. A
second patch will reduce HashMap lookups for positioning nodes as well.