Port YUV primitive to be a brush.
Also refactor and optimize the YUV shader implementation to do
less per-pixel work.

Support segment building for YUV primitive - this is a large
performance win for any YUV images with a complex clip on them.

This is part of the prep work to allow selection of whether
a stacking context is rasterized in local or screen-space.