Remove discard from radial gradient, apply tile/repeat on CPU.
Introduce a new method to decompose a PrimitiveInfo struct info
a (bounded) maximum number of instances based on a tile size and
repeat structure.

This allows us to remove the discard instruction from the radial
gradient shader, which means it doesn't disable early-Z.

This is a start point - in the future we'll want to expand this
to work on other gradients and images, and do a proper job
of handling edge cases with huge numbers of repetitions.

For now, this improves the common case of radial gradients, and
makes it easier to convert it to a brush shader.