Reduce the GPU cache size of pictures.
For the rare case of color matrix, store the color matrix as a
separate piece of GPU cache data.

This allows a smaller GPU cache size for the majority of picture
types. This is important to allow collapsing the brush_image and
ps_hw_composite code paths into one, which is the next step to
fixing up some drop-shadow bugs.