Remove aligned gradient shader.
Instead, push aligned gradients through the linear angle gradient
shader (for now).

This will simplify converting linear gradient code over to be
a brush shader. The optimization here was always a bit dubious
(it made more sense when we first started WR, when the angle
gradient shader was more expensive than it is now).

Instead, what we can do in the future if gradients ever show
up in the GPU profile is support vertex colors for the rectangle
shader and decompose simple aligned gradients into a series
of rectangle segments with different vertex colors.

The benefit of this is reduced shader changes and simpler batching.