Use FastTransform almost everywhere
FastTransform used to be called TransformOrOffset and this change uses it
in a lot of places that we used to use TypedTransform3D directly. This
helps to avoid the cost of expensive matrix math and allows us to more
liberally call things like `inverse()` on our transformations, since the
worst case is that we will be retrieving the already-cached inverted
matrix. This decreases the amount of time that FrameBuilder::build takes
in my local CPU profiles.