Properly handle optimized ClipChain in custom ClipChain nodes
When a node included in a ClipScrollNode ClipChain is optimized away, we need to
ensure that custom ClipChains that rely on that node are using the
original node and not one of the parents. In order to achieve this, we
store a copy of the original ClipChainNode in the ClipScrollNode and use
it when building custom ClipChains.

Fixes #2388.