Port drop-shadow and blur filter to use brush_image.
There are a couple of goals that will come out of this work:

 * We want to remove brush_picture shader altogether, collapsing
   it into brush_image. This is the first part of that. The final
   part is modifying box-shadows to draw with segments so that we
   don't have to tweak the UVs in the fragment shader. Fixing this
   will also simplify a number of other tasks related to existing
   box-shadow bugs.
 * Output descriptors of render tasks are written into the GPU cache
   as a normal ImageResource structure, the same format as items in
   the texture cache. The idea here is to allow the image drawing
   shader to not care whether the source is a render task or a
   normal static image. This simplifies the image drawing shader,
   and also helps get the code into a better position to cache
   the outputs of more render task types in the texture cache.

Adding a color to every texture cache item isn't ideal. I did some
profiling and it didn't make any noticeable performance impact.
In the future, if we notice a cost to it, we could add a separate
shader feature for A8 vs RGBA8 images, and have a different
texture cache item format for each of those shaders.