Optimization: draw solid, single color borders as rect + clip.
Due to various recent optimizations and features added for the
segment logic, it's now feasible to start drawing borders as
a solid rect + clip mask.

For now, we start with only the simplest and most common case,
borders with style: solid and a single color.

This optimization applies in a couple of ways:
 * Fewer shader switches and better batching. Specifically, we
   can often batch these simple borders into an existing solid
   rect batch. Additionally, the edge segments of borders are
   typically considered opaque segments, which are more efficient
   to batch, and help with early z-reject of other fragments.

 * Once we start to cache clip masks between elements, and also
   across frames / display lists, we'll automatically get some
   wins here by only rendering a border corner mask once where
   applicable. Many real world sites use a common border radius
   on all buttons, for example, so this can be a significant
   optimization in these cases.