Use brush_image for blurs, instead of 2d hardware composite.
Once this lands, we can fix drop-shadows by making use of this
functionality, and port the other remaining uses of ps_hardware_composite
to use brush_image. Then, ps_hardware_composite can be removed,
and everything except 3d plane-splits will then be drawn as
normal primitives.

Incidentally, this fixes the way we handle blurs when the source
picture is clipped by either the screen bounds or a clip-chain.
The code will now inflate the rendered portion of the source
render task correctly, so include enough pixels to get a correct
blur result at the edges.