Refactor the way picture and brush primitives are stored.
Previously, Picture was a separate primitive type, and it contained
a BrushPrimitive structure. Now there is a separate array of Picture
structures, and there is a BrushPrimitive::Picture type that refers
to a Picture by index.

This PR doesn't contain any functional changes - I think it will
be easier to get this reviewed and landed, and then work on the
functional changes as a separate patch.

The idea is to separate the concept of a Picture (a collection of
primitives, with children, optionally drawn into an intermediate
surface) from how that picture is drawn into the parent picture
(via BrushPrimitive::Picture).

What this means in the future is that a Picture can be referenced
by multiple Brush::Picture primitives and composited in a different
way. The primary use for this is for drop shadows - where we need
to draw the picture both as normal content, and also as an alpha
mask in a separate primitive, that has an offset / spread radius.

To achieve this, we'll move some fields currently on Picture (such
as filters, frame output) to the Brush::Picture primitive instance.

The benefits of such a change are:
 * We can reference a picture multiple times in the scene on a
   different primitive, which will allow us to fix outstanding
   drop-shadow bugs.
 * Since the Picture elements are now stored outside the primitive
   tree, we can start caching these across frames. This will form
   the basis of caching text-shadows, opacity filters etc.
 * We create fewer Picture instances (rather than one per filter
   stage as we do now). This helps debugging primarily.
 * Allowing a Picture to be referenced multiple times is more
   applicable to how SVG filter graphs work (when we implement those).

Since this is an intermediate patch, it introduces a bit of extra
indirection and duplication in places. I expect these to be resolved
by the follow up changes referenced above.