Early out from scene building if view is too large.
If the requested view size is larger than the maximum texture
size supported by the hardware, early exit and fail to render.

In the future, we could potentially render the scene in tiles,
however this is a quick fix for an existing crash.

Specifically, if Gecko is asked to open a very large window by
JS, the first frame has a very large window size. On subsequent
frames Gecko supplies a window size that is clamped to the
actual monitor resolution.

Fixes https://bugzilla.mozilla.org/show_bug.cgi?id=1441392 and
potentially https://bugzilla.mozilla.org/show_bug.cgi?id=1442921.