Clear the texture cache (and resource cache) before running each reftest in wrench.
The OSMesa rasterizer can give slightly different results for tests,
depending on where an image is located in the texture cache. This
attempts to mitigate that, by ensuring the texture cache is empty
at the start of each test.

This is an attempt to prevent the addition of a new reftest causing
unrelated test failures in later tests during a reftest run.