Handle box-shadow stretch mode per-axis.
This is a crash fix for box shadows that have one axis which is
very long, and one axis that is small.

It's also an optimization for box-shadows that have one long and
one short axis, since it allows drawing a minimal shadow rect in
these cases. This is handled by having a stretch mode per-axis
rather than one for the entire shadow primitive.

Fixes #2561.