Remove PictureKind enum.
Previously, we had separate Picture kinds for box shadows and
text shadows.

Now, box-shadows are drawn as clip sources, and text shadows are
drawn as a regular picture with a blur filter set as the picture
composite mode.

This PR should not contain any functional changes - it's just
re-arranging the code to no longer use PictureKind.