Make stacking context clipping explicit
Instead of relying on the combination of a stacking contexts which
render to intermediate surfaces and the PushStackingContext's clip node,
add an explicit clipping node parameter for stacking contexts. This is
necessary, because currently stacking context clipping is done by always
rendering to an intermediate surface. Once we support merging stacking
context and contained item clip chains we can remove this and the extra
parameter.