Fix AA on ellipse clip code under some edge cases.
If the AA range is > 1.0, due to a specific transform, the inner
pixels of a primitive that are not affected by the elliptical
radii in a clip mask may end up with a signed distance of 1.0,
which causes blending artifacts.

Instead, initialize the default value for signed distance to
the calculated AA range, which ensures we get a blend factor
of 1.0 in such cases.

Fixes #2576.