Ensure scene-only transactions still go through the scene builder thread.
If WR gets a transaction which only contains scene messages, and has the
flag to use the scene builder thread, it will not actually use the scene
builder thread, because the update_document function first drains all
the scene messages, and then checks to see if the transaction is empty.
In this scenario the transaction is empty (because all the scene
messages got drained) but the op.build flag is set to true. So instead
of delegating the scene build to the builder thread, the build is done
on the RB thread.

This patch modifies the condition so that all transactions that are
intended to be handled by the scene builder thread get handled there.