Add the SceneBuilderHooks object.
This replaces the render backend thread registration (which it turns out
is not that useful in it's current form) with a more comprehensive
SceneBuilderHooks object that is used to hook in at various points in
the scene builder thread.

This is needed because with async scene building, APZ will need to map
its "updater thread" to the WR scene builder thread, and so we need the
ability to run certain bits of code at certain points during the scene
builder thread's lifecycle.